refactor database store into separate folder
