+++ microcosm-fastapi/microcosm_fastapi/factories/postgres.py
@@ -1,8 +1,6 @@
 from microcosm_postgres.factories.engine import choose_args, choose_uri
 from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
 
-from microcosm_fastapi.context import SessionContext
-
 
 def make_engine(metadata, config):
     uri = choose_uri(metadata, config.postgres)
