+++ test-project/test_project/tests/test_route.py
@@ -4,7 +4,7 @@ from unittest.mock import ANY, patch
 import pytest
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
-from microcosm_fastapi.context import SessionContext, transaction
+from microcosm_fastapi.database.context import SessionContext, transaction
 from microcosm_postgres.identifiers import new_object_id
 from microcosm_postgres.operations import recreate_all
